Improved training stability with larger horizon_length
The motor dynamics are slow, requiring several steps for each pulicy
update. It takes around 2 seconds to switch the thruster input command
from -1.0 to 1.0 due to the maximum step implemented in the Kingfisher
Arduino board. For a step_dt of 0.05s (20 Hz), the rollout (or horizon
length) should be above 40 to cover enough experience for an entire
control cycle.

Additionally, the rewards have been adjusted, and the range for the
initial positions has been increased. In this version, the training is
stable for different dts, small reward differences, goal positions and
reinforcement learning libraries (rl_games and skrl)