Update simulation frequency to mach ECM
The Electronic Control Module in the kingfisher works at 50Hz and
updates its input command on each step. The environment was updated
accordingly and the force update was moved to apply_action to be changed
a bit faster. The step dt was decreased to 10Hz as the system dynamics
are slow